+++ packages/drizzle/src/factory/index.ts
@@ -2,6 +2,7 @@ import type { Table } from "drizzle-orm"
 import { MySqlDatabase, type MySqlTable } from "drizzle-orm/mysql-core"
 import { PgDatabase, type PgTable } from "drizzle-orm/pg-core"
 import type { BaseSQLiteDatabase, SQLiteTable } from "drizzle-orm/sqlite-core"
+import type { DrizzleFactoryInputVisibilityBehaviors } from "../types"
 import { DrizzleMySQLResolverFactory } from "./resolver-mysql"
 import { DrizzlePostgresResolverFactory } from "./resolver-postgres"
 import { DrizzleSQLiteResolverFactory } from "./resolver-sqlite"
@@ -12,7 +13,10 @@ export function drizzleResolverFactory<
   TTableName extends keyof NonNullable<TDatabase["_"]["schema"]>,
 >(
   db: TDatabase,
-  tableName: TTableName
+  tableName: TTableName,
+  options?: DrizzleFactoryInputVisibilityBehaviors<
+    NonNullable<TDatabase["_"]["fullSchema"]>[TTableName]
+  >
 ): DrizzleSQLiteResolverFactory<
   TDatabase,
   NonNullable<TDatabase["_"]["fullSchema"]>[TTableName]
@@ -20,14 +24,21 @@ export function drizzleResolverFactory<
 export function drizzleResolverFactory<
   TDatabase extends BaseSQLiteDatabase<any, any, any, any, any, any>,
   TTable extends SQLiteTable,
->(db: TDatabase, table: TTable): DrizzleSQLiteResolverFactory<TDatabase, TTable>
+>(
+  db: TDatabase,
+  table: TTable,
+  options?: DrizzleFactoryInputVisibilityBehaviors<TTable>
+): DrizzleSQLiteResolverFactory<TDatabase, TTable>
 
 export function drizzleResolverFactory<
   TDatabase extends PgDatabase<any, any, any, any, any>,
   TTableName extends keyof NonNullable<TDatabase["_"]["schema"]>,
 >(
   db: TDatabase,
-  tableName: TTableName
+  tableName: TTableName,
+  options?: DrizzleFactoryInputVisibilityBehaviors<
+    NonNullable<TDatabase["_"]["fullSchema"]>[TTableName]
+  >
 ): DrizzlePostgresResolverFactory<
   TDatabase,
   NonNullable<TDatabase["_"]["fullSchema"]>[TTableName]
@@ -37,7 +48,8 @@ export function drizzleResolverFactory<
   TTable extends PgTable,
 >(
   db: TDatabase,
-  table: TTable
+  table: TTable,
+  options?: DrizzleFactoryInputVisibilityBehaviors<TTable>
 ): DrizzlePostgresResolverFactory<TDatabase, TTable>
 
 export function drizzleResolverFactory<
@@ -45,7 +57,10 @@ export function drizzleResolverFactory<
   TTableName extends keyof NonNullable<TDatabase["_"]["schema"]>,
 >(
   db: TDatabase,
-  tableName: TTableName
+  tableName: TTableName,
+  options?: DrizzleFactoryInputVisibilityBehaviors<
+    NonNullable<TDatabase["_"]["fullSchema"]>[TTableName]
+  >
 ): DrizzleMySQLResolverFactory<
   TDatabase,
   NonNullable<TDatabase["_"]["fullSchema"]>[TTableName]
@@ -53,7 +68,11 @@ export function drizzleResolverFactory<
 export function drizzleResolverFactory<
   TDatabase extends MySqlDatabase<any, any, any, any, any, any>,
   TTable extends MySqlTable,
->(db: TDatabase, table: TTable): DrizzleMySQLResolverFactory<TDatabase, TTable>
+>(
+  db: TDatabase,
+  table: TTable,
+  options?: DrizzleFactoryInputVisibilityBehaviors<TTable>
+): DrizzleMySQLResolverFactory<TDatabase, TTable>
 
 export function drizzleResolverFactory(
   db: BaseDatabase,