+++ packages/drizzle/src/factory/resolver.ts
@@ -42,10 +42,16 @@ import {
   notLike,
   or,
 } from "drizzle-orm"
-import { DrizzleWeaver, type TableSilk } from ".."
+import { GraphQLInt, GraphQLNonNull } from "graphql"
+import {
+  type DrizzleResolverFactoryOptions,
+  DrizzleWeaver,
+  type TableSilk,
+} from ".."
 import { inArrayMultiple } from "../helper"
 import {
   type ColumnFilters,
+  type CountArgs,
   type DeleteArgs,
   DrizzleInputFactory,
   type FiltersCore,
@@ -58,6 +64,7 @@ import {
 import type {
   AnyQueryBuilder,
   BaseDatabase,
+  CountQuery,
   DeleteMutation,
   InferRelationTable,
   InferSelectArrayOptions,
@@ -86,9 +93,10 @@ export abstract class DrizzleResolverFactory<
   >
   public constructor(
     protected readonly db: TDatabase,
-    protected readonly table: TTable
+    protected readonly table: TTable,
+    protected readonly options?: DrizzleResolverFactoryOptions<TTable>
   ) {
-    this.inputFactory = new DrizzleInputFactory(table)
+    this.inputFactory = new DrizzleInputFactory(table, options)
     this.tableName = getTableName(table)
     const queryBuilder = this.db.query[
       this.tableName as keyof typeof this.db.query
@@ -171,6 +179,26 @@ export abstract class DrizzleResolverFactory<
       ...options,
       resolve: (opts) => {
         return queryBase.findFirst(opts) as any
+      },
+    } as QueryOptions<any, any>)
+  }
+
+  public countQuery<TInputI = CountArgs<TTable>>({
+    input,
+    ...options
+  }: GraphQLFieldOptions & {
+    input?: GraphQLSilk<CountArgs<TTable>, TInputI>
+    middlewares?: Middleware<CountQuery<TTable, TInputI>>[]
+  } = {}): CountQuery<TTable, TInputI> {
+    input ??= silk<CountArgs<TTable>, CountArgs<TTable>>(() =>
+      this.inputFactory.countArgs()
+    ) as GraphQLSilk<CountArgs<TTable>, TInputI>
+
+    return new QueryFactoryWithResolve(silk(new GraphQLNonNull(GraphQLInt)), {
+      input,
+      ...options,
+      resolve: (args) => {
+        return this.db.$count(this.table, this.extractFilters(args.where))
       },
     } as QueryOptions<any, any>)
   }
@@ -457,6 +485,7 @@ export abstract class DrizzleResolverFactory<
         ...fields,
         [name]: this.selectArrayQuery(),
         [`${name}Single`]: this.selectSingleQuery(),
+        [`${name}Count`]: this.countQuery(),
         [`insertInto${capitalize(name)}`]: this.insertArrayMutation(),
         [`insertInto${capitalize(name)}Single`]: this.insertSingleMutation(),
         [`update${capitalize(name)}`]: this.updateMutation(),